pyproject.toml
[project]
name = "pdntc-order"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.5.0",
    "pillow>=11.2.1",
    "pydantic-ai[logfire]>=0.1.3",
    "pydantic-evals>=0.1.3",
]

[dependency-groups]
dev = [
    "ruff>=0.11.6",
]
